Organize composer worklets
